requirements.txt
mcp>=0.5.0
httpx>=0.23.0
python-dotenv>=0.21.0
asyncio>=3.4.3
setuptools>=65.0.0
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.0
aiohttp>=3.8.4
GitPython>=3.1.40
markdown>=3.5.1

# Test dependencies
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.10.0

# Development dependencies
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.3.0 